Added optional timeout and max_retry args
Added an optional timeout argument for the initial connection, in seconds. Removed the hardcoded MAX_RETRY attribute and added an optional max_retry argument for the amount of times a request will retry after a timeout. Changed attempts and max_retry evaluation so it may work with a max_retry value of 1