Do not make assumption on Parts present for POST
The Redfish specification requires certain Parts to be present when uploading
to `MultipartHttpPushUri` (i.e., "UpdateParameters", "UpdateFile"),
but the `HttpPushUri` can have a vendor-specific message format.
Thus, to support both, remove the assumptions from the code with the
consequence that users need to pass the separate parts a bit more verbosely.

Signed-off-by: Max Crone <max.crone@brightcomputing.com>