Optional Connection upon Client Instantiation
This allows a user to create a client without _knowing_ the IP they are
connecting to up front. It gives the user the ability to create and pass
around a client object as needed. Additionally, the default behavior is
preserved.

* Added `check_connectivity` keyword arg to `RestClientBase`.
* Added `check_connectivity` keyword arg to `HTTPClient`.
* Added `check_connectivity` keyword arg to `redfish_client` function.
* Conditionally execute `get_root_object` during client instantiation
based on the value of `check_connectivity`.
* Catch the resultant `AttributeError` in the case that `self.root` was
not set due to `check_connectivity` being set to `False`.
* Small test refactor, added `setUp` method to `TestRedFishClient` class
to make commonly used attributes members of the test case.
* Added new test for this functionality.

Signed-off-by: Justin Palmer <jpavlav@gmail.com>